+++ R/estimate_ite_poisson.R
@@ -5,39 +5,44 @@
 #' Estimates the Individual Treatment Effect using Poisson Regression given a
 #' response vector, a treatment vector, and a features matrix.
 #'
-#' @param y the observed response vector
-#' @param z the treatment vector
-#' @param X the features matrix
-#' @param X_names the names of the covariates
-#' @param include_offset whether to include an offset (i.e. model outcome rate)
-#' or not (i.e. model outcome counts)
-#' @param offset_name the name of the covariate to use as offset (i.e. 'x1')
+#' @param y The observed response vector.
+#' @param z The treatment vector.
+#' @param X The features matrix.
+#' @param X_names The names of the covariates.
+#' @param offset Name of the covariate to use as offset (i.e. 'x1') to model the
+#' corresponding outcome rate. NULL to model drectly the outcome counts without
+#' offset.
 #'
 #' @return
-#' a vector of ITE estimates
+#' A vector of ITE estimates
 #'
 #' @keywords internal
 #'
 #'
-estimate_ite_poisson <- function(y, z, X, X_names, include_offset, offset_name){
-  if (include_offset) {
-    X_names = X_names[-which(X_names == offset_name)]
-    names(X)[names(X) == offset_name] <- 'offset'
-    y_treated <- data.frame(y = y[z==1])
-    X_treated <- as.data.frame(X[z==1,])
-    y_control <- data.frame(y = y[z==0])
-    X_control <- as.data.frame(X[z==0,])
-    data_treated <- cbind(y_treated, X_treated)
-    data_control <- cbind(y_control, X_control)
-    formula = as.formula(paste("y ~ ", paste(X_names, collapse= "+")))
-    temp1 <- stats::glm(formula,
-                        data = data_treated,
-                        offset = log(offset),
-                        family = stats::poisson(link = "log"))
-    temp0 <- stats::glm(formula,
-                        data = data_control,
-                        offset = log(offset),
-                        family = stats::poisson(link = "log"))
+estimate_ite_poisson <- function(y, z, X, X_names, offset){
+  if (!is.null(offset)) {
+    if (!(offset %in% X_names)){
+      stop("Offset varible is not observed. Please replace `offset` with an
+           observed varibale.")
+    } else {
+      X_names = X_names[-which(X_names == offset)]
+      names(X)[names(X) == offset] <- 'offset'
+      y_treated <- data.frame(y = y[z==1])
+      X_treated <- as.data.frame(X[z==1,])
+      y_control <- data.frame(y = y[z==0])
+      X_control <- as.data.frame(X[z==0,])
+      data_treated <- cbind(y_treated, X_treated)
+      data_control <- cbind(y_control, X_control)
+      formula = as.formula(paste("y ~ ", paste(X_names, collapse= "+")))
+      temp1 <- stats::glm(formula,
+                          data = data_treated,
+                          offset = log(offset),
+                          family = stats::poisson(link = "log"))
+      temp0 <- stats::glm(formula,
+                          data = data_control,
+                          offset = log(offset),
+                          family = stats::poisson(link = "log"))
+    }
   } else {
     y_treated <- data.frame(y = y[z==1])
     X_treated <- as.data.frame(X[z==1,])